% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumDE_Clean.R
\name{combData}
\alias{combData}
\title{Combine Data}
\usage{
combData(object, genes = c())
}
\arguments{
\item{object}{Seurat object to aggregate}

\item{genes}{List of genes to use, if not given use all genes}
}
\value{
data.all Pseudobulk matrix
}
\description{
Aggregate Data by labels in active.ident
}
