% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SumDE_Clean.R
\name{combDE}
\alias{combDE}
\title{Performs Pseudobulk DE}
\usage{
combDE(
  seur,
  cell_id = c(),
  kowt = "ko_vs_wt",
  combineOn = "orig.ident",
  method = "DESeq",
  form = "",
  minCells = 20,
  minBatches = 2,
  minReads = 10,
  mapIt = c()
)
}
\arguments{
\item{seur}{A seurat object}

\item{cell_id}{Array of cell types to perform DE on (assumes Cell Type info is in active.ident). If none given use all cells}

\item{kowt}{The column in the meta.data containing the condition of interest}

\item{combineOn}{Column in meta.data containing sample info for making Pseudobulk}

\item{method}{Method to use, DESeq for DESeq2, also can use EdgeR}

\item{form}{Formula to use. Can use to include extra covariate infomation. NOT IMPLEMENTED YET FOR edgeR.}

\item{minCells}{min number of cells per batch for batch to be included.}

\item{minBatches}{minimum number of batches per condition for analysis to run (if not met returns NULL).}

\item{minReads}{minimum number of reads in genes (requires this number of reads in at least 2 samples)}

\item{mapIt}{Not used anymore}
}
\value{
res DE results
}
\description{
Takes in a Seurat object and performs DE
}
