% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_Seurat.R
\name{dir10X}
\alias{dir10X}
\title{Make Seurat object}
\usage{
dir10X(
  seur = NULL,
  dat = NULL,
  minGenes = 500,
  regress = c("nCount_RNA"),
  species = "human"
)
}
\arguments{
\item{seur}{A seurat object, either this or dat required but not both}

\item{dat}{A count matrix, gene by cell, either this or seur required}

\item{minGenes}{Minimium number of genes per cell}

\item{regress}{List of parameters to regress out, pass c() if nothing to regress out}

\item{species}{Species (not currently used)}
}
\value{
seur A seurat object
}
\description{
Makes a Seurat object given a count matrix, or does downstream analysis on already existing Seurat object
}
