% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadCellBender.R
\name{loadCellBender}
\alias{loadCellBender}
\title{Load CellBender}
\usage{
loadCellBender(filename)
}
\arguments{
\item{filename}{File output by cellbender}
}
\value{
sparse.mat A sparse Matrix with expression counts
}
\description{
Loads the output of CellBender
}
