% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LoadCellBender.R
\name{loadCellBender_multi}
\alias{loadCellBender_multi}
\title{Load CellBender For Multiple Files}
\usage{
loadCellBender_multi(filenames)
}
\arguments{
\item{filenames}{List of files output by cellbender}
}
\value{
dat A sparse Matrix with expression counts
}
\description{
Loads the output of CellBender for multiple files
}
