% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunPropel.R
\name{runPropel}
\alias{runPropel}
\title{Run Propel}
\usage{
runPropel(seur, form, contrasts, samples, clusters)
}
\arguments{
\item{seur}{A seurat object to test}

\item{form}{A formula to use (for example ~0+condition+batch)}

\item{contrasts}{The contrast to use, for example conditionko-conditionwt}

\item{samples}{The column in meta.data with sample of origin information}

\item{clusters}{The column in meta.data with cluster/cell type information}
}
\value{
out Results of differential analysis
}
\description{
Run Propeller to test for chances in cluster abundance
}
