% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SharedVariable.R
\name{SharedVariable}
\alias{SharedVariable}
\title{Shared Variable Genes}
\usage{
SharedVariable(
  seur,
  x.low.cutoff = 1,
  x.high.cutoff = 5,
  minNum = 3,
  batch = "orig.ident",
  minCells = 10
)
}
\arguments{
\item{seur}{The Seurat object to run Variable Genes on}

\item{x.low.cutoff}{Parameter for variable genes}

\item{x.high.cutoff}{Parameter for variable genes}

\item{minNum}{Minimium number of batches/samples a variable gene must be found in to be used}

\item{batch}{Column in meta.data containing batch or sample information}

\item{minCells}{The min number of cells for a batch to be used}
}
\value{
A seurat object with new variable genes and active.ident set to batch
}
\description{
Calculaates variable genes shared between batches
}
